requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
torch
transformers
pillow
gunicorn
